+++ package.json
@@ -17,12 +17,13 @@
     "@radix-ui/react-slider": "^1.2.2",
     "@radix-ui/react-slot": "^1.1.1",
     "@shadcn/ui": "^0.0.4",
-    "@types/react": "^19.0.1",
-    "@types/react-dom": "^19.0.2",
+    "@types/react": "^18.3.16",
+    "@types/react-dom": "^18.3.5",
     "autoprefixer": "^10.4.20",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "date-fns": "^3.6.0",
+    "firebase": "^11.1.0",
     "framer-motion": "^11.14.4",
     "lucide-react": "^0.468.0",
     "postcss": "^8.4.49",
