+++ src/pages/index.astro
@@ -1,5 +1,5 @@
 ---
-import AchievementManagementPage from "@/pages/_AchievementManagementPage";
+import AchievementManagementPage from "@/pages/AchievementManagementPage";
 import Layout from "@/layouts/Layout.astro";
 ---
